+++ sdk/src/androidTest/java/ly/count/android/sdk/CountlyTests.java
@@ -161,7 +161,10 @@ public class CountlyTests {
         mUninitedCountly.init(new CountlyConfig(getContext(), "appkey", "http://test.count.ly").setDeviceId(null));
     }
 
-    @Test
+    /**
+     * This test not fail anymore because we are now not throwing an exception when device ID is empty.
+     */
+    @Test(expected = AssertionError.class)
     public void testInit_emptyDeviceID() {
         try {
             mUninitedCountly.init(new CountlyConfig(getContext(), "appkey", "http://test.count.ly").setDeviceId(""));
@@ -169,6 +172,15 @@ public class CountlyTests {
         } catch (IllegalArgumentException ignored) {
             // success!
         }
+    }
+
+    @Test
+    public void testInit_emptyDeviceID_sdkGenerated() {
+        mCountly.halt(); // to reset the state
+        mUninitedCountly.init(TestUtils.createBaseConfig().setDeviceId(""));
+
+        // this should be an SDK generated device ID
+        Assert.assertEquals(DeviceIdType.OPEN_UDID, mUninitedCountly.deviceId().getType());
     }
 
     @Test
@@ -389,7 +401,7 @@ public class CountlyTests {
         assertEquals(0, mCountly.getActivityCount());
         assertTrue(mCountly.getPrevSessionDurationStartTime() > 0);
         verify(requestQueueProvider).endSession(0);
-        verify(requestQueueProvider, times(0)).recordEvents(anyString());
+        verify(requestQueueProvider, times(1)).recordEvents(anyString()); // not 0 anymore, it will send orientation event
     }
 
     /**
@@ -522,7 +534,7 @@ public class CountlyTests {
         mCountly.onTimer();
 
         verify(requestQueueProvider).updateSession(0);
-        verify(requestQueueProvider, times(0)).recordEvents(anyString());
+        verify(requestQueueProvider, times(1)).recordEvents(anyString()); // not 0 anymore, it will send orientation event
     }
 
     @Test