+++ sdk/src/androidTest/java/ly/count/android/sdk/ModuleViewsTests.java
@@ -1652,7 +1652,8 @@ public class ModuleViewsTests {
             "arrDO", new JSONArray(arrDO),
             "arrBO", new JSONArray(arrBO),
             "arrIO", new JSONArray(arrIO),
-            "arrObjStr", new JSONArray(arrObjStr)
+            "arrObjStr", new JSONArray(arrObjStr),
+            "arrObj", new JSONArray(arrObj)
         ));
 
         // Validate the recorded event with expected segmentation
@@ -1706,7 +1707,8 @@ public class ModuleViewsTests {
             "arrL", arrL,
             "arrD", arrD,
             "arrBO", arrBO,
-            "arrIO", arrIO
+            "arrIO", arrIO,
+            "arrObj", arrObj
         ));
 
         // Validate the recorded event with expected segmentation