+++ sdk/src/androidTest/java/ly/count/android/sdk/TestUtils.java
@@ -42,7 +42,7 @@ public class TestUtils {
     public final static String commonAppKey = "appkey";
     public final static String commonDeviceId = "1234";
     public final static String SDK_NAME = "java-native-android";
-    public final static String SDK_VERSION = "24.7.0";
+    public final static String SDK_VERSION = "24.7.1";
     public static final int MAX_THREAD_COUNT_PER_STACK_TRACE = 50;
 
     public static class Activity2 extends Activity {
@@ -510,6 +510,14 @@ public class TestUtils {
         }
 
         return resultMapArray;
+    }
+
+    protected static void removeRequestContains(String search) {
+        for (String request : getCountyStore().getRequests()) {
+            if (request.contains(search)) {
+                getCountyStore().removeRequest(request);
+            }
+        }
     }
 
     protected static Map<String, Object> map(Object... args) {